email-service/requirements.txt
Flask==2.2.3
Flask-CORS==3.0.10
python-dotenv==0.21.1